python (FFI) example を削除 (#432)
