--- example/pyo3/.gitignore
@@ -1,8 +1,0 @@
-# OpenJTalk-dictionary's dir
-open_jtalk_dic_utf_8-*
-
-# shared library
-*.so
-*.so.*
-*.dylib
-*.dll
