+++ example/python/run.py
@@ -1,53 +1,78 @@
-import argparse
+import dataclasses
+import json
+import logging
+from argparse import ArgumentParser
+from pathlib import Path
+from typing import Tuple
 
-import core
+import voicevox_core
+from voicevox_core import AccelerationMode, AudioQuery, VoicevoxCore
+
+SPEAKER_ID = 0
 
 
-def run(
-    use_gpu: bool,
-    text: str,
-    speaker_id: int,
-    cpu_num_threads: int,
-    openjtalk_dict: str,
-    output: str,
-) -> None:
-    # コアの初期化
-    core.initialize(use_gpu, cpu_num_threads, load_all_models=False)
+def main() -> None:
+    logging.basicConfig(
+        format="[%(levelname)s] %(filename)s: %(message)s", level="DEBUG"
+    )
+    logger = logging.getLogger(__name__)
 
-    # openjtalk辞書のロード
-    core.voicevox_load_openjtalk_dict(openjtalk_dict)
+    (acceleration_mode, open_jtalk_dict_dir, text, out) = parse_args()
 
-    # 話者のロード
-    core.load_model(speaker_id)
+    logger.debug("%s", f"{voicevox_core.METAS=}")
+    logger.debug("%s", f"{voicevox_core.SUPPORTED_DEVICES=}")
 
-    # AudioQueryの生成
-    audio_query = core.voicevox_audio_query(text, speaker_id)
+    logger.info("%s", f"Initializing ({acceleration_mode=}, {open_jtalk_dict_dir=})")
+    core = VoicevoxCore(
+        acceleration_mode=acceleration_mode, open_jtalk_dict_dir=open_jtalk_dict_dir
+    )
 
-    # 音声合成
-    wavefmt = core.voicevox_synthesis(audio_query, speaker_id)
+    logger.debug("%s", f"{core.is_gpu_mode=}")
 
-    # 保存
-    with open(output, "wb") as f:
-        f.write(wavefmt)
+    logger.info("%s", f"Loading model {SPEAKER_ID}")
+    core.load_model(SPEAKER_ID)
 
-    core.finalize()
+    logger.debug("%s", f"{core.is_model_loaded(0)=}")
+
+    logger.info("%s", f"Creating an AudioQuery from {text!r}")
+    audio_query = core.audio_query(text, SPEAKER_ID)
+
+    logger.info("%s", f"Synthesizing with {display_as_json(audio_query)}")
+    wav = core.synthesis(audio_query, SPEAKER_ID)
+
+    out.write_bytes(wav)
+    logger.info("%s", f"Wrote `{out}`")
+
+
+def parse_args() -> Tuple[AccelerationMode, Path, str, Path]:
+    argparser = ArgumentParser()
+    argparser.add_argument(
+        "--mode",
+        default="AUTO",
+        type=AccelerationMode,
+        help='モード ("AUTO", "CPU", "GPU")',
+    )
+    argparser.add_argument(
+        "open_jtalk_dict_dir",
+        type=Path,
+        help="Open JTalkの辞書ディレクトリ",
+    )
+    argparser.add_argument(
+        "text",
+        help="読み上げさせたい文章",
+    )
+    argparser.add_argument(
+        "out",
+        type=Path,
+        help="出力wavファイルのパス",
+    )
+    args = argparser.parse_args()
+    return (args.mode, args.open_jtalk_dict_dir, args.text, args.out)
+
+
+def display_as_json(audio_query: AudioQuery) -> str:
+    return json.dumps(dataclasses.asdict(audio_query), ensure_ascii=False)
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--use_gpu", action="store_true")
-    parser.add_argument("--text", required=True)
-    parser.add_argument("--speaker_id", type=int, required=True)
-    parser.add_argument("--cpu_num_threads", type=int, default=0)
-    parser.add_argument(
-        "--openjtalk_dict",
-        type=str,
-        default="voicevox_core/open_jtalk_dic_utf_8-1.11"
-    )
-    parser.add_argument("--output", type=str)
-    
-    args = parser.parse_args()
-    if args.output is None:
-        args.output = f"{args.text}-{args.speaker_id}.wav"
-
-    run(**vars(args))
+    main()
